Morning warning cleanup: 2 unused imports and an unneccesary cast.

+++ src/gwt/client/shop/ShopPage.java
@@ -41,7 +41,7 @@ public class ShopPage extends Page
         } else if (action.equals(FAVORITES)) {
             // get the member id from args, if no member id, use the current user
             int memberId = args.get(1, CShop.getMemberId());
-            byte itemType = (byte) args.get(2, (int) Item.NOT_A_TYPE);
+            byte itemType = (byte) args.get(2, Item.NOT_A_TYPE);
             String[] prefixArgs = new String[] { FAVORITES, String.valueOf(memberId) };
             // TODO get member's name (passed from the profile page? loaded from service?) from id
             // and display it in the page title
